Remote poller route added
